Hide away RequestAPI type
