+++ packages/expo-server/src/runtime/scope.ts
@@ -1,4 +1,9 @@
-import { RequestAPI } from './api';
+export interface RequestAPI {
+  origin?: string;
+  environment?: string | null;
+  waitUntil?(promise: Promise<unknown>): void;
+  deferTask?(fn: () => Promise<unknown>): void;
+}
 
 export interface ScopeDefinition<Scope extends RequestAPI = any> {
   getStore(): Scope | undefined;
